[REFACTOR] 글로벌 공통 모듈 정리 및 예외 처리 리팩토링
