+++ src/main/kotlin/io/gijung/aichatbot/auth/application/exception/DuplicateEmailException.kt
@@ -1,6 +1,6 @@
 package io.gijung.aichatbot.auth.application.exception
 
-import io.gijung.aichatbot.exception.exception.AbstractCustomException
+import io.gijung.aichatbot.global.error.exception.AbstractCustomException
 
 class DuplicateEmailException(message: String) :
     AbstractCustomException(error = UserAccountError.DUPLICATE_EMAIL, message = message)
