+++ src/main/kotlin/io/gijung/aichatbot/auth/application/exception/NotFoundEmailException.kt
@@ -1,6 +1,6 @@
 package io.gijung.aichatbot.auth.application.exception
 
-import io.gijung.aichatbot.exception.exception.AbstractCustomException
+import io.gijung.aichatbot.global.error.exception.AbstractCustomException
 
 class NotFoundEmailException (message: String) :
     AbstractCustomException(error = UserAccountError.INVALID_CREDENTIALS, message = message)
