+++ src/main/kotlin/io/gijung/aichatbot/auth/domain/exception/EmailValidationException.kt
@@ -1,7 +1,7 @@
 package io.gijung.aichatbot.auth.domain.exception
 
-import io.gijung.aichatbot.exception.error.CustomError
-import io.gijung.aichatbot.exception.exception.AbstractCustomException
+import io.gijung.aichatbot.global.error.code.CustomError
+import io.gijung.aichatbot.global.error.exception.AbstractCustomException
 
 class EmailValidationException(error: CustomError, message: String) :
     AbstractCustomException(error = error, message = message)
